pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-checkpoint"
version = "1.0.0"
description = "A comprehensive MCP configuration scanner with client-aware security analysis"
authors = [
    { name = "Aira Security", email = "founders@airasecurity.ai" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.8"
license-files = ["LICEN[CS]E*"]
dependencies = [
    "fastmcp>=2.13.0",
    "pyyaml>=6.0.2",
    "thefuzz>=0.22.1",
    "rich>=14.1.0",
    "transformers==4.57.1",
    "torch==2.9.0",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[project.scripts]
mcp-checkpoint = "mcp_checkpoint.cli:main"

[project.urls]
Homepage = "https://github.com/aira-security/mcp-checkpoint"
Repository = "https://github.com/aira-security/mcp-checkpoint"
Issues = "https://github.com/aira-security/mcp-checkpoint/issues"

[tool.setuptools]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mcp_checkpoint*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true